urban_vision/requirements.txt
torch
torchvision
transformers
numpy
opencv-python
scipy
Pillow
requests